feat(erDiagram): adds Examples component
